+++ tests/test_complex_columns.py
@@ -3,7 +3,7 @@ import logging
 import pytest
 from pyspark.sql import functions as F
 from pyspark.sql.types import StructType, StructField, IntegerType, StringType, FloatType, ArrayType, MapType, \
-    BinaryType, LongType
+    BinaryType, LongType, DateType
 
 import dbldatagen as dg
 
@@ -21,6 +21,15 @@ class TestComplexColumns:
     dfTestData = None
     row_count = 1000
     column_count = 10
+
+    @staticmethod
+    def getFieldType(schema, fieldName):
+        fields = [fld for fld in schema.fields if fld.name == fieldName]
+
+        if fields is not None and len(fields) > 0:
+            return fields[0].dataType
+        else:
+            return None
 
     @pytest.mark.parametrize("complexFieldType, expectedType, invalidValueCondition",
                              [("array<int>", ArrayType(IntegerType()), "complex_field is not Null"),
@@ -396,3 +405,356 @@ class TestComplexColumns:
 
         for r in rows:
             assert r['test'] is not None
+
+    def test_inferred_column_types_disallowed1(self, setupLogging):
+        with pytest.raises(ValueError):
+            column_count = 10
+            data_rows = 10 * 1000
+            df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                       .withIdOutput()
+                       .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                                   numColumns=column_count, structType="array")
+                       .withColumn("code1", "integer", minValue=100, maxValue=200)
+                       .withColumn("code2", "integer", minValue=0, maxValue=10)
+                       .withColumn("code3", dg.INFER_DATATYPE, values=['a', 'b', 'c'])
+                       .withColumn("code4", StringType(), values=['a', 'b', 'c'], random=True)
+                       .withColumn("code5", StringType(), values=['a', 'b', 'c'], random=True, weights=[9, 1, 1])
+                       )
+
+            df = df_spec.build()
+
+            assert df is not None
+
+    def test_inferred_disallowed_with_schema(self):
+        """Test use of schema"""
+        schema = StructType([
+            StructField("region_id", IntegerType(), True),
+            StructField("region_cd", StringType(), True),
+            StructField("c", StringType(), True),
+            StructField("c1", StringType(), True),
+            StructField("state1", StringType(), True),
+            StructField("state2", StringType(), True),
+            StructField("st_desc", StringType(), True),
+
+        ])
+
+        testDataSpec = (dg.DataGenerator(spark, name="test_data_set1", rows=10000)
+                        .withSchema(schema)
+                        )
+
+        with pytest.raises(ValueError):
+            testDataSpec2 = testDataSpec.withColumnSpecs(matchTypes=[dg.INFER_DATATYPE], minValue=0, maxValue=100)
+            df = testDataSpec2.build()
+            df.show()
+
+    def test_inferred_column_basic(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="code1 + code2")
+                   .withColumn("code7", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   )
+
+        columnSpec1 = df_spec.getColumnSpec("code1")
+        assert columnSpec1.inferDatatype is False
+
+        columnSpec5 = df_spec.getColumnSpec("code5")
+        assert columnSpec5.inferDatatype is True
+
+        columnSpec6 = df_spec.getColumnSpec("code6")
+        assert columnSpec6.inferDatatype is True
+
+        columnSpec7 = df_spec.getColumnSpec("code7")
+        assert columnSpec7.inferDatatype is True
+
+    def test_inferred_column_validate_types(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="code1 + code2")
+                   .withColumn("code7", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "code5")
+        assert type1 == DateType()
+
+        type2 = self.getFieldType(df.schema, "code6")
+        assert type2 == IntegerType()
+
+        type3 = self.getFieldType(df.schema, "code7")
+        assert type3 == StringType()
+
+    def test_inferred_column_structs1(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withColumn("struct1", dg.INFER_DATATYPE, expr="named_struct('a', code1, 'b', code2)")
+                   .withColumn("struct2", dg.INFER_DATATYPE, expr="named_struct('a', code5, 'b', code6)")
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        expectedType = StructType([StructField('a', IntegerType()), StructField('b', IntegerType())])
+        assert type1 == expectedType
+
+        type2 = self.getFieldType(df.schema, "struct2")
+        expectedType2 = StructType([StructField('a', DateType(), False), StructField('b', StringType())])
+        assert type2 == expectedType2
+
+    def test_inferred_column_structs2(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withColumn("struct1", dg.INFER_DATATYPE, expr="named_struct('a', code1, 'b', code2)")
+                   .withColumn("struct2", dg.INFER_DATATYPE, expr="named_struct('a', code5, 'b', code6)")
+                   .withColumn("struct3", dg.INFER_DATATYPE, expr="named_struct('a', struct1, 'b', struct2)")
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        assert type1 == StructType([StructField('a', IntegerType()), StructField('b', IntegerType())])
+        type2 = self.getFieldType(df.schema, "struct2")
+        assert type2 == StructType([StructField('a', DateType(), False), StructField('b', StringType())])
+        type3 = self.getFieldType(df.schema, "struct3")
+        assert type3 == StructType(
+            [StructField('a', StructType([StructField('a', IntegerType()), StructField('b', IntegerType())]), False),
+             StructField('b', StructType([StructField('a', DateType(), False), StructField('b', StringType())]), False)]
+        )
+
+    def test_with_struct_column1(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withStructColumn("struct1", fields=[('a', 'code1'), ('b', 'code2')])
+                   .withStructColumn("struct2", fields=[('a', 'code5'), ('b', 'code6')])
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        assert type1 == StructType([StructField('a', IntegerType()), StructField('b', IntegerType())])
+        type2 = self.getFieldType(df.schema, "struct2")
+        assert type2 == StructType([StructField('a', DateType(), False), StructField('b', StringType())])
+
+    def test_with_struct_column2(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withStructColumn("struct1", fields=['code1', 'code2'])
+                   .withStructColumn("struct2", fields=['code5', 'code6'])
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        assert type1 == StructType([StructField('code1', IntegerType()), StructField('code2', IntegerType())])
+        type2 = self.getFieldType(df.schema, "struct2")
+        assert type2 == StructType([StructField('code5', DateType(), False), StructField('code6', StringType())])
+
+    def test_with_json_struct_column(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withStructColumn("struct1", fields=['code1', 'code2'], asJson=True)
+                   .withStructColumn("struct2", fields=['code5', 'code6'], asJson=True)
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        print("type1", type1)
+        assert type1 == StringType()
+        type2 = self.getFieldType(df.schema, "struct2")
+        assert type2 == StringType()
+
+    def test_with_json_struct_column2(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withStructColumn("struct1", fields={'codes': ["code6", "code6"]}, asJson=True)
+                   .withStructColumn("struct2", fields=['code5', 'code6'], asJson=True)
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        print("type1", type1)
+        assert type1 == StringType()
+
+        type2 = df_spec.inferredSchema["struct1"].dataType
+        assert type2 == StringType()
+
+    def test_with_struct_column3(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withStructColumn("struct1", fields=[('a', 'code1'), ('b', 'code2')])
+                   .withStructColumn("struct2", fields=[('a', 'code5'), ('b', 'code6')])
+                   .withStructColumn("struct3", fields=[('a', 'struct1'), ('b', 'struct2')])
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        assert type1 == StructType([StructField('a', IntegerType()), StructField('b', IntegerType())])
+        type2 = self.getFieldType(df.schema, "struct2")
+        assert type2 == StructType([StructField('a', DateType(), False), StructField('b', StringType())])
+        type3 = self.getFieldType(df.schema, "struct3")
+        assert type3 == StructType(
+            [StructField('a', StructType([StructField('a', IntegerType()), StructField('b', IntegerType())]), False),
+             StructField('b', StructType([StructField('a', DateType(), False), StructField('b', StringType())]),
+                         False)])
+
+    def test_with_struct_column4(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+        df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                   .withIdOutput()
+                   .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                               numColumns=column_count, structType="array")
+                   .withColumn("code1", "integer", minValue=100, maxValue=200)
+                   .withColumn("code2", "integer", minValue=0, maxValue=10)
+                   .withColumn("code3", "string", values=['one', 'two', 'three'])
+                   .withColumn("code4", "string", values=['one', 'two', 'three'])
+                   .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                   .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                   .withStructColumn("struct1", fields=[('a', 'code1'), ('b', 'code2')])
+                   .withStructColumn("struct2", fields=[('a', 'code5'), ('b', 'code6')])
+                   .withStructColumn("struct3",
+                                     fields={'a': {'a': 'code1', 'b': 'code2'}, 'b': {'a': 'code5', 'b': 'code6'}})
+                   )
+
+        df = df_spec.build()
+
+        type1 = self.getFieldType(df.schema, "struct1")
+        assert type1 == StructType([StructField('a', IntegerType()), StructField('b', IntegerType())])
+        type2 = self.getFieldType(df.schema, "struct2")
+        assert type2 == StructType([StructField('a', DateType(), False), StructField('b', StringType())])
+        type3 = self.getFieldType(df.schema, "struct3")
+        assert type3 == StructType(
+            [StructField('a', StructType([StructField('a', IntegerType()), StructField('b', IntegerType())]), False),
+             StructField('b', StructType([StructField('a', DateType(), False), StructField('b', StringType())]),
+                         False)])
+
+    def test_with_struct_column_err1(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+
+        with pytest.raises(ValueError):
+            df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                       .withIdOutput()
+                       .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                                   numColumns=column_count, structType="array")
+                       .withColumn("code1", "integer", minValue=100, maxValue=200)
+                       .withColumn("code2", "integer", minValue=0, maxValue=10)
+                       .withColumn("code3", "string", values=['one', 'two', 'three'])
+                       .withColumn("code4", "string", values=['one', 'two', 'three'])
+                       .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                       .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                       .withStructColumn("struct1", fields={'BAD_FIELD': 45})
+                       )
+
+            df = df_spec.build()
+
+    def test_with_struct_column_err2(self, setupLogging):
+        column_count = 10
+        data_rows = 10 * 1000
+
+        with pytest.raises(Exception):
+            df_spec = (dg.DataGenerator(spark, name="test_data_set1", rows=data_rows)
+                       .withIdOutput()
+                       .withColumn("r", FloatType(), expr="floor(rand() * 350) * (86400 + 3600)",
+                                   numColumns=column_count, structType="array")
+                       .withColumn("code1", "integer", minValue=100, maxValue=200)
+                       .withColumn("code2", "integer", minValue=0, maxValue=10)
+                       .withColumn("code3", "string", values=['one', 'two', 'three'])
+                       .withColumn("code4", "string", values=['one', 'two', 'three'])
+                       .withColumn("code5", dg.INFER_DATATYPE, expr="current_date()")
+                       .withColumn("code6", dg.INFER_DATATYPE, expr="concat(code3, code4)")
+                       .withStructColumn("struct1", fields=23)
+                       )
+
+            df = df_spec.build()